Added apps for all planned features
This is just structural stuff, obviously.  I think the way we're
breaking things down is:

 * frio: news, schedule, comms (twitter)
 * rzn:  servers
 * fs:   screens

The FAQ and nearby apps are nice-to-haves, rather than must-haves, and
are currently unassigned.  Tournaments is another nice-to-have (we can
just link directly to the challonge) so is unassigned, and comms will
basically just poll Twitter.