Switches to Heroku/12-factor style development
Most settings are now derived -- either from the environment (and
therefore .env thanks to honcho etc.), or by using __file__ to figure
out whereabouts the project is located.  This should mean that both
developers should be able to clone the project, setup the environment,
and run it with no configuration (yay!).